requirements.txt
# Core dependencies
Flask>=3.0.0
Flask-Cors>=5.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.23.0

# API clients
openai>=1.0.0
alpaca-trade-api>=3.0.0
IEXfinance>=0.5.0

# Data handling
requests>=2.28.0
websocket-client>=1.5.0
yfinance>=0.2.30

# Notifications
twilio>=9.0.0
plyer>=2.0.0  # For desktop notifications
pygame>=2.6.0  # For voice notifications
pyttsx3>=2.90  # For fallback voice notifications
hume>=0.7.0   # For Hume AI voice notifications

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0

# Machine Learning
scikit-learn>=1.2.0

# Testing
pytest>=7.3.0

# Utilities
tqdm>=4.64.0
pytz>=2023.3

# For Windows environments
pywin32>=303; platform_system=="Windows"
